requirements.txt
python-fasthtml
uvicorn
fastapi